scraper/requirements.txt
httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.41.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
